Revert "add `labels` module; revise `selector_load`; +10% perf (#46)" (#49)
PR #46 has caused performance degradation on the GKE cluster, according
to the results of Locust and Hey load testing. Surprisingly, `cargo bench`
gives contradictory result --- it reports 8–10% *improve* of performance.

I trust the results returned by Locust and Hey tools more than those
returned by `cargo bench` --- the Locust script exercises a set of
different PromQL queries, while `cargo bench` only checks one.

It's sad to see this PR go. (It does improve the code, and I worked hard
on it.) Still, the sacrifice of code quality in favour of performance
is the sacrifice we are willing to make. So be it.

- This reverts commit 6fe6e81c594623a0291f5ddc0a53ea369693c86f.
- This reverts commit 502d260f96e317803f5e99baff9be5c3e952f525.
- This reverts commit 82c74bd17508dd5fe1b1ef4d676601b7fbf68022.